GCP/requirements.txt
functions-framework==3.*
google-cloud-storage
Pillow
tensorflow==2.14.0
numpy==1.26.2
requests
flask